refactor: move total-related component to more appropriate package
